feat(routes): added route system
